+++ FireRender.Maya.Src/FireRenderPBRMaterial.cpp
@@ -189,7 +189,7 @@ namespace FireMaya
 
 		MDistance::Unit sceneUnits = MDistance::uiUnit();
 		MDistance distance(1.0, sceneUnits);
-		float scale_multiplier = distance.asMeters();
+		float scale_multiplier = (float) distance.asMeters();
 
 		auto ms = scope.MaterialSystem();
